+++ src/components/App.js
@@ -1,15 +1,12 @@
 import React from "react";
 import "../sass/style.css";
 import Footer from "./Footer";
-import BackToTop from "./Backtotop";
 import Navbarhere from "./Menu";
-
 
 const App = () => {
   return (
     <>
       <Navbarhere />
-      <BackToTop />
       <Footer />
     </>
   );
